requirements.txt
playwright
pytest
requests
python-dotenv
allure-pytest
